Create emailListTest.js, update get requests at frontend

+++ frontend/src/components/adminCollection/toDoCollection/ToDos/ToDos.jsx
@@ -46,17 +46,19 @@ const ToDos = ({ toDos, resetRes, resUpdate, setResUpdate, resDelete, setResDele
   }
 
   useEffect(() => {
-    fetch(`http://localhost:3001/to-do-list/${deleteById}`, {
-      method: 'DELETE',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-        'authorization': localStorage.getItem('token')
-      }
-    }).then(res => res.json())
-      .then(res => setResDelete(resDelete + 1))
-      .catch(err => setResDelete(false))
-      .finally(() => resetRes())
+    if (deleteById) {
+      fetch(`http://localhost:3001/to-do-list/${deleteById}`, {
+        method: 'DELETE',
+        headers: {
+          'Accept': 'application/json',
+          'Content-Type': 'application/json',
+          'authorization': localStorage.getItem('token')
+        }
+      }).then(res => res.json())
+        .then(res => setResDelete(resDelete + 1))
+        .catch(err => setResDelete(false))
+        .finally(() => resetRes())
+    }
   }, [changeDelete])
 
   return (
